Implement/derive common traits
From the rust API guidelines [1], this resolves C-COMMON-TRAITS,
C-DEBUG, and C-DEBUG-NONEMPTY. The criterion used to decide whether
or not to implement a trait was whether the implementation makes
logical sense, not whether it was likely to be useful. I advise that,
in the future, all additions to this library should implement these
traits as well.

[1] https://rust-lang-nursery.github.io/api-guidelines/about.html