Make usage reporting safe
HAL_Report expects a null-terminated string to be passed as feature.
This ensures that a valid UTF-8 null-terminated string is passed in,
making report_usage_extras a safe wrapper around HAL_Report.

This also adds a report_usage_context function to allow for the
otherwise lost ability to pass in a null pointer as feature.

Also return the return value of HAL_Report, in case anyone actually
finds it useful.